Add wrapper factory
