+++ tests/MiddlewareFactoryTest.php
@@ -21,6 +21,7 @@ use Yiisoft\Middleware\Dispatcher\Tests\Support\UseParamsMiddleware;
 use Yiisoft\Middleware\Dispatcher\Tests\Support\InvalidController;
 use Yiisoft\Middleware\Dispatcher\Tests\Support\TestController;
 use Yiisoft\Middleware\Dispatcher\Tests\Support\TestMiddleware;
+use Yiisoft\Middleware\Dispatcher\WrapperFactory;
 use Yiisoft\Test\Support\Container\SimpleContainer;
 
 final class MiddlewareFactoryTest extends TestCase
@@ -256,10 +257,10 @@ final class MiddlewareFactoryTest extends TestCase
     private function getMiddlewareFactory(ContainerInterface $container = null): MiddlewareFactoryInterface
     {
         if ($container !== null) {
-            return new MiddlewareFactory($container);
+            return new MiddlewareFactory($container, new WrapperFactory($container));
         }
 
-        return new MiddlewareFactory($this->getContainer());
+        return new MiddlewareFactory($this->getContainer(), new WrapperFactory($this->getContainer()));
     }
 
     private function getContainer(array $instances = []): ContainerInterface
